Finalizando: Separar generador de fechas
